requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
gdown==5.2.0
h11==0.16.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.2.1
pydantic_core==2.33.2
pydantic==2.11.7
PySocks==1.7.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.35.0
